DWAA support - functions moved to plugin_tools.py
Obsolete files removed
