+++ pype/lib/__init__.py
@@ -33,7 +33,16 @@ from .applications import (
     _subprocess
 )
 
-from .plugin_tools import filter_pyblish_plugins, source_hash
+from .plugin_tools import (
+    filter_pyblish_plugins,
+    source_hash,
+    get_unique_layer_name,
+    get_background_layers,
+    oiio_supported,
+    decompress,
+    get_decompress_dir,
+    should_decompress
+)
 
 from .path_tools import (
     version_up,
@@ -72,6 +81,12 @@ __all__ = [
     "ApplicationAction",
 
     "filter_pyblish_plugins",
+    "get_unique_layer_name",
+    "get_background_layers",
+    "oiio_supported",
+    "decompress",
+    "get_decompress_dir",
+    "should_decompress",
 
     "version_up",
     "get_version_from_path",
