+++ pype/plugins/global/publish/extract_jpeg.py
@@ -3,7 +3,8 @@ import os
 import pyblish.api
 import pype.api
 import pype.lib
-from pype.plugins import lib as plugins_lib
+from pype.lib import oiio_supported, should_decompress, \
+    get_decompress_dir, decompress
 import shutil
 
 
@@ -24,10 +25,10 @@ class ExtractJpegEXR(pyblish.api.InstancePlugin):
         if 'crypto' in instance.data['subset']:
             return
 
-        decompress = False
+        do_decompress = False
         # ffmpeg doesn't support multipart exrs, use oiiotool if available
         if instance.data.get("multipartExr") is True:
-            if not plugins_lib.oiio_supported():
+            if not oiio_supported():
                 return
 
         # Skip review when requested.
@@ -61,11 +62,11 @@ class ExtractJpegEXR(pyblish.api.InstancePlugin):
             self.log.info("input {}".format(full_input_path))
 
             decompressed_dir = ''
-            decompress = plugins_lib.should_decompress(full_input_path)
-            if decompress:
-                decompressed_dir = plugins_lib.get_decompress_dir()
+            do_decompress = should_decompress(full_input_path)
+            if do_decompress:
+                decompressed_dir = get_decompress_dir()
 
-                plugins_lib.decompress(
+                decompress(
                     decompressed_dir,
                     full_input_path)
                 # input path changed, 'decompressed' added
@@ -124,7 +125,7 @@ class ExtractJpegEXR(pyblish.api.InstancePlugin):
             self.log.debug("Adding: {}".format(representation))
             representations_new.append(representation)
 
-            if decompress and os.path.exists(decompressed_dir):
+            if do_decompress and os.path.exists(decompressed_dir):
                 shutil.rmtree(decompressed_dir)
 
         instance.data["representations"] = representations_new
